Attempt at rewriting the [tab] shortcode to only enqueue its scripts when called.
https://github.com/INN/umbrella-jjie/commit/0b503660901ff5920fd3fcea47f1fef513158a98#diff-4a8c90930ba6b0891dbd2868b303565aL7 removed the scripts enqueued for the shortcode, because I thought that the scripts were only being enqueued on page 103871, and page 103871 in production was /hub/, which didn't have the [tab] shortcode.

So that was wrong, but the scripts should only be enqueued on the shortcode's usage, not at all times for all pages that are children of the /hub/ page.

This is a first pass at a fix for that, but it's broken right now because the scripts aren't being enqueued.